+++ lib/mix/tasks/ref_inspector/download.ex
@@ -37,7 +37,8 @@ defmodule Mix.Tasks.RefInspector.Download do
   ]
 
   def run(args) do
-    :ok = start_app()
+    :ok = start_app(args)
+    :ok = Config.init_env()
 
     {opts, _argv, _errors} = OptionParser.parse(args, @cli_options)
 
@@ -84,14 +85,18 @@ defmodule Mix.Tasks.RefInspector.Download do
     end
   end
 
-  defp start_app do
-    project = Mix.Project.config()
+  defp start_app(args) do
+    _ = Mix.Task.run("loadpaths", args)
 
-    if project[:app] not in Application.started_applications() do
+    unless "--no-compile" in args do
+      _ = Mix.Project.compile(args)
+    end
+
+    if :ref_inspector not in Application.started_applications() do
       startup_silent = Application.get_env(:ref_inspector, :startup_silent, false)
       :ok = Application.put_env(:ref_inspector, :startup_silent, true)
 
-      _ = Mix.Task.run("app.start")
+      _ = Application.ensure_all_started(:ref_inspector)
 
       :ok = Application.put_env(:ref_inspector, :startup_silent, startup_silent)
     end